debugging: pressure forces
- debugging pressure forces
- Convert all to double for accuracy

+++ SPH/header/Particle.h
@@ -3,25 +3,25 @@
 #include <glm\vec3.hpp>
 #include "Constant.h"
 
-using glm::vec3;
+using glm::dvec3;
 using glm::ivec3;
 
 class Particle
 {
 public:
 	Particle() {
-		position = vec3(0.0, 0.0, 0.0);
+		position = dvec3(0.0, 0.0, 0.0);
 		//default color blue
-		color = vec3(0.0, 0.0, 1.0);
+		color = dvec3(0.0, 0.0, 1.0);
 		reCalculateGridCell();
 	};
-	Particle(vec3& pos, int _id) {
-		position = vec3(pos.x, pos.y, pos.z);
+	Particle(dvec3& pos, int _id) {
+		position = dvec3(pos.x, pos.y, pos.z);
 		id = _id;
 		
-		velocity = vec3(0.0f);
-		acceleration = vec3(0.0f);
-		prevAcceleration = vec3(0.0f);
+		velocity = dvec3(0.0);
+		acceleration = dvec3(0.0);
+		prevAcceleration = dvec3(0.0);
 
 		density = REST_DENSITY;
 		pressure = 0.0;
@@ -37,16 +37,16 @@ public:
 
 public:
 	int id;
-	vec3 position;
+	dvec3 position;
 	ivec3 cellPosition;
-	vec3 color;
+	dvec3 color;
 	//sph field
-	vec3 velocity;
-	vec3 acceleration;
-	vec3 prevAcceleration;//for leapfrog
+	dvec3 velocity;
+	dvec3 acceleration;
+	dvec3 prevAcceleration;//for leapfrog
 	
-	float density;
-	float pressure;
+	double density;
+	double pressure;
 };
 
 #endif
